Split map and supplier buildings tabs
