+++ src/router/index.js
@@ -4,6 +4,7 @@ import IslandInfoPage from '@/views/IslandInfoPage.vue';
 import PopulationPage from '@/views/PopulationPage.vue'
 import TreasuryPage from '@/views/TreasuryPage.vue'
 import BuildingsPage from '@/views/BuildingsPage.vue'
+import YieldBuildingsPage from '@/views/YieldBuildingsPage.vue'
 import ManufacturesPage from '@/views/ManufacturesPage.vue'
 import HomeView from '@/views/HomeView.vue';
 import LoginView from '@/views/LoginView.vue';
@@ -34,6 +35,7 @@ const routes = [
             { path: 'population', name: 'population', component: PopulationPage },
             { path: 'treasury', name: 'treasury', component: TreasuryPage },
             { path: 'manufactures', name: 'manufactures', component: ManufacturesPage },
+            { path: 'yield-buildings', name: 'yield-buildings', component: YieldBuildingsPage },
         ]
     },
     { path: '/ships', component: ShipsView, meta: { requiresAuth: false } },
